refactor ticketCreate
