+++ src/containers/Transactions/Simple/index.js
@@ -11,7 +11,6 @@ import DepositPreauth from './DepositPreauth'
 import EnableAmendment from './EnableAmendment'
 import UNLModify from './UNLModify'
 import AccountDelete from './AccountDelete'
-import TicketCreate from './TicketCreate'
 import { transactionTypes } from '../../shared/components/Transaction'
 import { useLanguage } from '../../shared/hooks'
 
@@ -49,8 +48,6 @@ const Simple = (props) => {
       return <UNLModify t={t} language={language} data={data} />
     case 'AccountDelete':
       return <AccountDelete t={t} language={language} data={data} />
-    case 'TicketCreate':
-      return <TicketCreate t={t} language={language} data={data} />
     default:
       // Some transactions do not have simple views.
       return (
