+++ src/containers/shared/components/Transaction/TicketCreate/index.ts
@@ -1,0 +1,11 @@
+import { TransactionMapping } from '../types'
+
+import { Simple } from './Simple'
+import { parser } from './parser'
+import { TableDetail } from './TableDetail'
+
+export const TicketCreateTransaction: TransactionMapping = {
+  Simple,
+  TableDetail,
+  parser,
+}
