+++ src/containers/shared/components/Transaction/TicketCreate/parser.ts
@@ -1,4 +1,4 @@
-module.exports = (tx) => {
+export function parser(tx: any) {
   const ticketCount = tx.TicketCount
   const ticketSequence = tx.TicketSequence
   return {
