+++ src/containers/shared/components/Transaction/TicketCreate/test/mock_data/TicketCreate.json
