+++ src/containers/shared/components/TxDetails.tsx
@@ -238,19 +238,6 @@ const TxDetails = (props: Props) => {
       </div>
     )
   }
-
-  function renderTicketCreate(): ReactElement {
-    const { instructions } = props
-    const { ticketCount } = instructions
-    return (
-      <div className="ticketCreate">
-        <span className="label">{t('ticket_count')}:</span>
-        <span> </span>
-        <span>{ticketCount}</span>
-      </div>
-    )
-  }
-
   const { type = '', instructions } = props
   const functionMap: { [key: string]: () => ReactElement | null } = {
     renderAccountSet,
@@ -258,7 +245,6 @@ const TxDetails = (props: Props) => {
     renderPayment,
     renderPaymentChannelCreate,
     renderPaymentChannelClaim,
-    renderTicketCreate,
   }
 
   // Locate the component for detail row that is unique per TransactionType.
